+++ src/contact/index.ts
@@ -1,0 +1,4 @@
+export * from './entities/contactCreate';
+export * from './entities/contactList';
+export * from './entities/contact';
+export * from './contactClient';