+++ src/publicApiClient.ts
@@ -4,6 +4,7 @@ import { LicensesClient } from './licenses/licensesClient';
 import { SubscriptionsClient } from './subscriptions/subscriptionsClient';
 import { CustomersClient } from './customers/customersClient';
 import { OrdersClient } from './orders';
+import { ContactClient } from './contact/contactClient';
 
 /**
  * Public API Client class, should be the main entry point for your calls
@@ -69,6 +70,16 @@ export class PublicApiClient extends AbstractClient {
    */
   public getOrdersClient(): OrdersClient {
     return new OrdersClient(this.client)
+      .setUrl(this.url)
+      .setApiKey(this.apiKey);
+  }
+
+  /**
+   * Creates a new {@link ContactClient} instance and returns it
+   * @returns {@link ContactClient}
+   */
+  public getContactClient(): ContactClient {
+    return new ContactClient(this.client)
       .setUrl(this.url)
       .setApiKey(this.apiKey);
   }