+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.0] - 2022-05-17
+
+### Changed
+
+- Adding 4 new EndPoints of Contact Information:
+  - Create Contact
+  - Get a Contact
+  - Get list of Contact
+  - Update a Contact
+
 ## [3.5.0] - 2022-05-02
 
 ### Changed